split edns_parser tests
